requirements.txt
# ============================================
# Property Comparison Tool - All Dependencies
# ============================================

# Core Dependencies
python-dotenv>=1.0.0
pydantic>=2.0.0
pyyaml>=6.0.0

# ============================================
# LangGraph & LangChain (for graph-based pipeline)
# ============================================
langgraph>=0.0.40
langchain>=0.1.0
langchain-core>=0.1.23
langchain-openai>=0.0.5
beautifulsoup4>=4.12.0
langchain-anthropic>=0.1.0

# ============================================
# AI/LLM Providers
# ============================================
openai>=1.3.0
anthropic>=0.7.0

# ============================================
# NLP & Text Processing
# ============================================
scikit-learn>=1.3.0
# spacy>=3.7.0  # Optional - uncomment if needed

# ============================================
# Reporting & Templates
# ============================================
jinja2>=3.1.0
markdown>=3.5.0

# ============================================
# CLI & Utilities
# ============================================
click>=8.1.0
rich>=13.0.0
colorama>=0.4.6

# ============================================
# Data Processing
# ============================================
beautifulsoup4>=4.12.0
lxml>=4.9.0

# ============================================
# Web Scraping
# ============================================
firecrawl-py>=0.0.16  # Clean web scraping API for LLM-ready markdown

# Future alternatives (not needed with Firecrawl):
# playwright>=1.40.0
# selenium>=4.15.0
